Show deprecation warning when stateful lambdas are used as kernels (#21885)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/21885

If a kernel is defined as a stateful lambda

    static auto registry = torch::RegisterOperators().op("my::op", [some_closure] (Tensor a) {...});

this can have very unexpected behavior when kernels are instantiated. There is no guarantee that the state is kept.

In the options based API, state is already disallowed:

    // this is a compiler error
    static auto registry = torch::RegisterOperators().op("my::op", torch::RegisterOperators::options().kernel([some_closure] (Tensor a) {...}));

but we can't disallow it in the non-options-based API for backwards compatibility reasons.

We can, however, show a deprecation warning. This is what this diff introduces.

Differential Revision: D15867089

fbshipit-source-id: 300fa4772fad8e7d177eb7cb910063d360537a4a